Committed 2021/06/04

+++ Python Machine Learning by Example/Chapter5 - Logistic Regression/Implement logistic regression using TensorFlow_3.py
@@ -71,7 +71,7 @@ for step, (batch_x, batch_y) in enumerate(train_data.take(training_steps), 1):
         print("step: %i, loss: %f" % (step, loss))
 
 
-X_test_enc = np.array(X_test_enc)
+X_test_enc = np.array(X_test_enc, dtype=tf.float32)
 
 logits = tf.add(tf.matmul(X_test_enc, weight), bias)[:, 0]
 prediction = tf.nn.sigmoid(logits)
